# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Face Swap API
  description: |
    Backend for the iOS Face Swap app. Uses Novita.ai for face merge, video face swap, and motion-controlled video generation.

    **Key points for iOS:**
    - All request bodies are **JSON with base64** (no multipart). Encode images/videos as base64 strings.
    - **Sync:** `POST /api/v1/merge-face` returns the result image in the response.
    - **Direct result:** `POST /api/v1/video-merge-face` and `POST /api/v1/motion-control` submit the job to Novita and return the task result directly in the response.
    - Base URL: your Railway deployment (e.g. `https://your-app.up.railway.app`).
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: /
    description: Current host (use your Railway URL, e.g. https://your-app.up.railway.app)
...
paths:
...
  /api/v1/merge-face:
    post:
      tags: [Merge Face]
      summary: Merge face (sync)
      description: |
        Combine two faces into one image. **Synchronous** — the response contains the result image as base64.
        Use this when you have a face image and a target image and want a single merged image back immediately.
      operationId: mergeFace
      parameters:
        - name: X-Device-ID
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MergeFaceRequest'
            example:
              face_image_base64: "<base64-encoded-face-image>"
              image_base64: "<base64-encoded-target-image>"
              image_type: png
      responses:
        '200':
          description: Success — merged image returned as base64
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MergeFaceResponse'
        '400':
          description: Invalid request (e.g. missing or invalid fields)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '402':
          description: Insufficient credits
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit exceeded (5 requests per minute per device)
        '500':
          description: Server error (e.g. failed to record task)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '502':
          description: Upstream error (Novita or internal)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/video-merge-face:
    post:
      tags: [Video Merge Face]
      summary: Video face swap
      description: |
        Replace the face in a video with a face from an image. Submits the job to Novita and returns the task result directly.
        The `video_asset_id` must be a Novita asset ID (upload your video to Novita assets first).
      operationId: videoMergeFace
      parameters:
        - name: X-Device-ID
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VideoMergeFaceRequest'
            example:
              video_asset_id: "<novita-asset-id>"
              face_image_base64: "<base64-encoded-face-image>"
      responses:
        '200':
          description: Task result from Novita
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NovitaTaskResult'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '402':
          description: Insufficient credits
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit exceeded (5 requests per minute per device)
        '500':
          description: Server error (e.g. failed to record task)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '502':
          description: Upstream error (Novita or upload failed)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/motion-control:
    post:
      tags: [Motion Control]
      summary: Motion control
      description: |
        Apply motion from a reference video to a reference image (Kling V2.6 Pro). Submits the job to Novita and returns the task result directly.
      operationId: motionControl
      parameters:
        - name: X-Device-ID
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MotionControlRequest'
            example:
              reference_image_base64: "<base64-encoded-image>"
              reference_video_base64: "<base64-encoded-video>"
              positive_prompt: ""
              negative_prompt: ""
              character_orientation: video
              keep_original_sound: false
      responses:
        '200':
          description: Task result from Novita
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NovitaTaskResult'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '402':
          description: Insufficient credits
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit exceeded (5 requests per minute per device)
        '500':
          description: Server error (e.g. failed to record task)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '502':
          description: Upstream error (Novita)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/v1/categories:
    get:
      tags: [Categories]
      summary: List categories
      description: Returns categories ordered by sort_order (ascending). Optionally filter by type to get categories for a specific tab.
      operationId: listCategories
      parameters:
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [image, video, motion_control]
          description: Filter categories by tab type
      responses:
        '200':
          description: List of categories
          content:
            application/json:
              schema:
                type: object
                properties:
                  categories:
                    type: array
                    items:
                      $ref: '#/components/schemas/Category'
        '400':
          description: Invalid type parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags: [Categories]
      summary: Create category
      description: Creates a new template category.
      operationId: createCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, type]
              properties:
                name:
                  type: string
                type:
                  type: string
                  enum: [image, video, motion_control]
                sort_order:
                  type: integer
                  default: 0
      responses:
        '201':
          description: Category created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/categories/{category_id}:
    patch:
      tags: [Categories]
      summary: Update category
      description: Updates an existing category. All fields are optional.
      operationId: updateCategory
      parameters:
        - name: category_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                type:
                  type: string
                  enum: [image, video, motion_control]
                sort_order:
                  type: integer
      responses:
        '200':
          description: Updated category
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Category not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags: [Categories]
      summary: Delete category
      description: Deletes a category and all its templates (cascade).
      operationId: deleteCategory
      parameters:
        - name: category_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Deleted
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message
    MergeFaceRequest:
      type: object
      required:
        - face_image_base64
        - image_base64
      properties:
        face_image_base64:
          type: string
          description: Base64-encoded face image
        image_base64:
          type: string
          description: Base64-encoded target image to merge the face into
        image_type:
          type: string
          default: png
          enum: [png, jpg]
          description: Output (and input) image type
    MergeFaceResponse:
      type: object
      properties:
        image_base64:
          type: string
          description: Base64-encoded merged result image
        image_type:
          type: string
          example: png
    VideoMergeFaceRequest:
      type: object
      required:
        - video_asset_id
        - face_image_base64
      properties:
        video_asset_id:
          type: string
          description: Novita asset ID for the target video (upload to Novita assets first)
        face_image_base64:
          type: string
          description: Base64-encoded face image
        template_id:
          type: string
          format: uuid
          nullable: true
          description: Optional template ID
    MotionControlRequest:
      type: object
      required:
        - reference_image_base64
        - reference_video_base64
      properties:
        reference_image_base64:
          type: string
          description: Base64-encoded reference image (max 10MB)
        reference_video_base64:
          type: string
          description: Base64-encoded reference video — motion source (max 50MB)
        positive_prompt:
          type: string
          default: ""
          description: Up to 2500 characters
        negative_prompt:
          type: string
          default: ""
          description: Up to 2500 characters
        character_orientation:
          type: string
          default: video
          description: Orientation of the character in the output
        keep_original_sound:
          type: boolean
          default: false
        template_id:
          type: string
          format: uuid
          nullable: true
          description: Optional template ID
    NovitaTaskResult:
      type: object
      properties:
        task:
          type: object
          properties:
            task_id:
              type: string
            status:
              type: string
              example: TASK_STATUS_SUCCEED
              enum:
                - TASK_STATUS_QUEUED
                - TASK_STATUS_PROCESSING
                - TASK_STATUS_SUCCEED
                - TASK_STATUS_FAILED
            output:
              type: object
              description: Task output — structure varies by endpoint (e.g. video URL for face swap/motion control)
              nullable: true
    Category:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        type:
          type: string
          enum: [image, video, motion_control]
        sort_order:
          type: integer
...